Refactor: Move agent logic to core and improve imports
This commit reorganizes the project structure by moving the agent-related logic to the `core/agent.py` file. It also improves import handling for optional dependencies like `SetupRunner` and removes unnecessary `sys.path` manipulation. The `agent.py` file now serves as a backward-compatible alias.
