+++ src/ai_sidekick_for_splunk/core/__init__.py
@@ -5,19 +5,23 @@ This module contains the foundational classes and utilities for building
 modular, extensible Splunk AI agents using Google ADK.
 """
 
-from .base_agent import BaseAgent
-from .base_tool import BaseTool
+from .base_agent import BaseAgent, AgentMetadata
+from .base_tool import BaseTool, ToolMetadata
 from .config import Config
 from .discovery import ComponentDiscovery
-from .orchestrator import SplunkOrchestrator
-from .registry import AgentRegistry, ToolRegistry
+from .orchestrator import SplunkOrchestrator, create_agent
+from .registry import AgentRegistry, RegistryManager, ToolRegistry
 
 __all__ = [
     "SplunkOrchestrator",
+    "create_agent",
     "BaseAgent",
+    "AgentMetadata",
     "BaseTool",
+    "ToolMetadata",
     "Config",
     "AgentRegistry",
     "ToolRegistry",
+    "RegistryManager",
     "ComponentDiscovery",
 ]
