+++ src/ai_sidekick_for_splunk/services/__init__.py
@@ -5,6 +5,10 @@ This package contains services for managing sessions, execution,
 and other infrastructure concerns for the AI Sidekick for Splunk agent.
 """
 
-from .setup_runner import SetupRunner
-
-__all__ = ["SetupRunner"]
+# Import with graceful fallback for optional dependencies
+try:
+    from .setup_runner import SetupRunner  # noqa: F401
+    __all__ = ["SetupRunner"]
+except ImportError:
+    # Google ADK not available - services not available
+    __all__ = []
